- Hoàn thành tính năng sửa hợp đồng - Thêm validate vào cho hợp đồng

+++ src/main/java/com/mobifone/transmission/dto/inf/SiteViewDTO.java
@@ -3,13 +3,15 @@ package com.mobifone.transmission.dto.inf;
 import com.mobifone.transmission.model.SiteOwner;
 import com.mobifone.transmission.model.TransmissionOwner;
 
+import java.math.BigDecimal;
+
 public interface SiteViewDTO {
     Long getId();
     String getSiteId();
     String getSiteId2();
     String getSiteName();
-    Float getLatitude();
-    Float getLongitude();
+    BigDecimal getLatitude();
+    BigDecimal getLongitude();
     String getAddress();
     SiteOwnerWithNameDTO getSiteOwner();
     interface SiteOwnerWithNameDTO {
